FIX: Homing issue, if we are sitting exactly on the bed at 0,0,0
There is a issue were if we are sitting dead center on the Z0.0 position then we will fail to home by about 6mm. I am not sure if this is a setup issue in the config or not. I don't think so as most of the settings are default for the kossel